linux: Add CentOS 7 based container for x86_64
Thanks to the SCL, we can get the GCC 8 toolchain,
and GCC 9 should be made available once 9.3 is released.

The setup is cleaner than on Ubuntu and we have an even
older glibc version for compatibility. The drawback is
that there is no SCL for CentOS 7 i686, and Mono upstream
also doesn't provide packages for it (though we could build
them from the src.rpm).

And of course, Ubuntu 14.04 is now EOL so it makes sense to
make our builds on a still supported release.