JavaScript: Install Emscripten SDK 1.39.9 (Mono) and 2.0.10 (Standard).
Mono needs to stay on 1.39.9 for compatibility with Mono 6.12.x, but for
standard builds the latest is best, and 2.0.10+ is needed for GDNative support.

None is sourced by default, they will be sourced by the build script.

We use a copy of the same Git clone to ensure that `emsdk_env.sh` is
compatible, otherwise we end up with a dirty state when sourcing both.