Cleanup and improve README
Removed old instructions about Ubuntu 18.04 as there's no good reason
to use such an old OS to build those containers, and we're not testing
it so chances are that the containers may not build/work properly there.

It saves a lot of trouble to just use latest Fedora (or possibly RHEL/CentOS)
to build those containers.